Move input parsing out of the requests package to input package
- Update other code accordingly
