+++ impl/pcc/plugin.go
@@ -8,7 +8,7 @@ import (
 	"github.com/gemfire/cloudcache-management-cf-plugin/domain"
 	"github.com/gemfire/cloudcache-management-cf-plugin/impl/common"
 	"github.com/gemfire/cloudcache-management-cf-plugin/util"
-	"github.com/gemfire/cloudcache-management-cf-plugin/util/requests"
+	"github.com/gemfire/cloudcache-management-cf-plugin/util/input"
 )
 
 // BasicPlugin declares the dataset that commands work on
@@ -28,7 +28,7 @@ func (c *BasicPlugin) Run(cliConnection plugin.CliConnection, args []string) {
 		return
 	}
 	var err error
-	c.commandData.Target, c.commandData.UserCommand = requests.GetTargetAndClusterCommand(args)
+	c.commandData.Target, c.commandData.UserCommand = input.GetTargetAndClusterCommand(args)
 	if c.commandData.UserCommand.Command == "" {
 		fmt.Println(util.GenericErrorMessage, err.Error())
 		os.Exit(1)
