+++ util/output/output_suite_test.go
@@ -9,5 +9,5 @@ import (
 
 func TestFormat(t *testing.T) {
 	RegisterFailHandler(Fail)
-	RunSpecs(t, "Format Suite")
+	RunSpecs(t, "Output Suite")
 }
