+++ util/requests/helpers_test.go
@@ -3,7 +3,7 @@ package requests_test
 import (
 	"os"
 
-	"github.com/gemfire/cloudcache-management-cf-plugin/util/requests"
+	"github.com/gemfire/cloudcache-management-cf-plugin/util/input"
 	. "github.com/onsi/ginkgo"
 	. "github.com/onsi/gomega"
 )
@@ -22,21 +22,21 @@ var _ = Describe("Helpers", func() {
 		Context("with no target in environment", func() {
 			It("returns no target, no command", func() {
 				args = []string{"program"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal(""))
 				Expect(userCommand.Command).To(Equal(""))
 			})
 
 			It("returns target but no command", func() {
 				args = []string{"program", "target"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal(""))
 			})
 
 			It("returns target and multiple word command", func() {
 				args = []string{"program", "target", "list", "members"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(0))
@@ -44,7 +44,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target, multiple word command and options ", func() {
 				args = []string{"program", "target", "list", "members", "-h"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(1))
@@ -54,7 +54,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target, multiple word command and option with values ", func() {
 				args = []string{"program", "target", "list", "members", "-t", "abc"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(1))
@@ -64,7 +64,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target, multiple word command, option without value and option with values ", func() {
 				args = []string{"program", "target", "list", "members", "-h", "-t", "abc"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(2))
@@ -75,7 +75,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target, multiple word command, option with value and option without values ", func() {
 				args = []string{"program", "target", "list", "members", "-t", "abc", "-h"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(2))
@@ -95,21 +95,21 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target but no command", func() {
 				args = []string{"program", "target"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal(""))
 			})
 
 			It("returns target and command", func() {
 				args = []string{"program", "command"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("command"))
 			})
 
 			It("returns target and multiple word command", func() {
 				args = []string{"program", "target", "list", "members"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(0))
@@ -117,7 +117,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target and multiple word command", func() {
 				args = []string{"program", "list", "members"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(0))
@@ -125,7 +125,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target, multiple word command and options ", func() {
 				args = []string{"program", "target", "list", "members", "-h"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(1))
@@ -135,7 +135,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target, multiple word command and options ", func() {
 				args = []string{"program", "list", "members", "-h"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(1))
@@ -145,7 +145,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target, multiple word command and option with values ", func() {
 				args = []string{"program", "list", "members", "-t", "abc"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(1))
@@ -155,7 +155,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target, multiple word command, option without value and option with values ", func() {
 				args = []string{"program", "list", "members", "-h", "-t", "abc"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(2))
@@ -166,7 +166,7 @@ var _ = Describe("Helpers", func() {
 
 			It("returns target, multiple word command, option with value and option without values ", func() {
 				args = []string{"program", "list", "members", "-t", "abc", "-h"}
-				target, userCommand := requests.GetTargetAndClusterCommand(args)
+				target, userCommand := input.GetTargetAndClusterCommand(args)
 				Expect(target).To(Equal("target"))
 				Expect(userCommand.Command).To(Equal("list members"))
 				Expect(len(userCommand.Parameters)).To(Equal(2))
